+++ CHANGELOG.md
@@ -14,7 +14,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 
 ### Fixed
 - Nav plugin: the `order` option is not applied.
-- Updated dependencies: `std`.
+- Updated dependencies: `std`, `postcss`.
 
 ## [2.2.2] - 2024-06-21
 ### Fixed